+++ R/apply-enzyme.R
@@ -191,13 +191,13 @@ apply_enzyme <- function(glycans, enzyme, return_list = NULL) {
 #'   of the rule on the corresponding glycan.
 #' @noRd
 .match_rule <- function(glycans, rule) {
-  res <- glymotif::match_motif(
+  res <- .match_motif(
     glycans,
     rule$acceptor,
     alignment = rule$acceptor_alignment
   )
   if (length(rule$rejects) > 0) {
-    rej_match_res <- glymotif::match_motifs(
+    rej_match_res <- .match_motifs(
       glycans,
       rule$rejects,
       alignments = rule$acceptor_alignment